+++ frontend/src/main.tsx
@@ -4,6 +4,7 @@ import { BrowserRouter } from 'react-router-dom'
 import App from './App'
 import './styles.css'
 import { LangContext, useLangState } from './i18n'
+import { ErrorBoundary } from './components/ErrorBoundary'
 
 function LangProvider({ children }: { children: React.ReactNode }) {
   const value = useLangState()
@@ -12,10 +13,12 @@ function LangProvider({ children }: { children: React.ReactNode }) {
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <BrowserRouter>
-      <LangProvider>
-        <App />
-      </LangProvider>
-    </BrowserRouter>
+    <ErrorBoundary>
+      <BrowserRouter>
+        <LangProvider>
+          <App />
+        </LangProvider>
+      </BrowserRouter>
+    </ErrorBoundary>
   </React.StrictMode>
 )